refactor prompt-page state to zustand
