+++ package.json
@@ -90,7 +90,8 @@
     "tsd": "^0.32.0",
     "vaul": "^1.1.2",
     "vivus": "^0.4.6",
-    "zod": "^3.24.2"
+    "zod": "^3.24.2",
+    "zustand": "^5.0.7"
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4.0.9",
